chore(component): remove previous Select implementation
