+++ packages/headless/src/index.ts
@@ -4,5 +4,4 @@ export * from './components/drawer';
 export * from './components/tabs/tabs';
 export * from './components/toggle/toggle';
 export * from './components/tooltip/tooltip';
-export * from './components/select/select';
-export * as AltSelect from './components/select/alt-select';
+export * as Select from './components/select/select';
