+++ packages/website/src/routes/docs/daisy/select/index.tsx
@@ -1,5 +1,5 @@
 import { component$, useId, useStyles$ } from '@builder.io/qwik';
-import { AltSelect } from '@qwik-ui/headless';
+import { Select } from '@qwik-ui/headless';
 import styles from './select.css?inline';
 
 export default component$(() => {
@@ -7,22 +7,20 @@ export default component$(() => {
   return (
     <>
       <h2>This is the documentation for the Select</h2>
-      <AltSelect.Root>
-        <AltSelect.Label class="mr-2">
+      <Select.Root>
+        <Select.Label class="mr-2">
           Alternate Daisy Select implementation
-        </AltSelect.Label>
-        <AltSelect.Trigger class="select-caret inline-flex items-center justify-between w-full h-12 max-w-xs px-4 text-sm font-semibold leading-loose border rounded-lg cursor-pointer border-base-content border-opacity-20 bg-base-100 shrink-0">
-          <AltSelect.Value placeholder="Select an option! ⚡" />
-        </AltSelect.Trigger>
-        <AltSelect.ListBox class="w-full max-w-xs border border-base-content border-opacity-20 bg-base-100 rounded-[0.25rem] py-1">
-          <AltSelect.Option
+        </Select.Label>
+        <Select.Trigger class="select-caret inline-flex items-center justify-between w-full h-12 max-w-xs px-4 text-sm font-semibold leading-loose border rounded-lg cursor-pointer border-base-content border-opacity-20 bg-base-100 shrink-0">
+          <Select.Value placeholder="Select an option! ⚡" />
+        </Select.Trigger>
+        <Select.ListBox class="w-full max-w-xs border border-base-content border-opacity-20 bg-base-100 rounded-[0.25rem] py-1">
+          <Select.Option
             value="🚀 Qwik"
             class="px-1 leading-loose hover:bg-gray-300"
           />
-          <AltSelect.Group class="">
-            <AltSelect.Label class="block px-1 leading-loose">
-              Fruits
-            </AltSelect.Label>
+          <Select.Group class="">
+            <Select.Label class="block px-1 leading-loose">Fruits</Select.Label>
             {[
               { value: '🍎 Apple', disabled: false },
               { value: '🍌 Banana', disabled: false },
@@ -30,7 +28,7 @@ export default component$(() => {
               { value: '🐲 Dragonfruit', disabled: true },
             ].map((option) => {
               return (
-                <AltSelect.Option
+                <Select.Option
                   key={useId()}
                   value={option.value}
                   disabled={option.disabled}
@@ -38,9 +36,9 @@ export default component$(() => {
                 />
               );
             })}
-          </AltSelect.Group>
-        </AltSelect.ListBox>
-      </AltSelect.Root>
+          </Select.Group>
+        </Select.ListBox>
+      </Select.Root>
     </>
   );
 });
